requirements.txt
imageio==2.37.0
matplotlib==3.10.0
numpy==2.2.2
pandas==2.2.3
pillow==11.1.0
scikit-learn==1.6.1
scipy==1.15.1
